chore: rename DOStateStore to CloudflareStateStore; rename "alchemy-state" worker to "alchemy-state-service"
