+++ alchemy/src/state/cloudflare-state-store.ts
@@ -17,10 +17,10 @@ import { logger } from "../util/logger.ts";
 import { memoize } from "../util/memoize.ts";
 import { StateStoreProxy } from "./proxy.ts";
 
-export interface DOStateStoreOptions extends CloudflareApiOptions {
+export interface CloudflareStateStoreOptions extends CloudflareApiOptions {
   /**
    * The name of the script to use for the state store.
-   * @default "alchemy-state"
+   * @default "alchemy-state-service"
    */
   scriptName?: string;
   /**
@@ -42,10 +42,10 @@ export interface DOStateStoreOptions extends CloudflareApiOptions {
  *
  * @see {@link https://alchemy.run/guides/do-state-store DOStateStore}
  */
-export class DOStateStore extends StateStoreProxy {
+export class CloudflareStateStore extends StateStoreProxy {
   constructor(
     scope: Scope,
-    private readonly options: DOStateStoreOptions = {},
+    private readonly options: CloudflareStateStoreOptions = {},
   ) {
     super(scope);
   }
@@ -71,7 +71,7 @@ export class DOStateStore extends StateStoreProxy {
       });
       if (!response.headers.get("Content-Type")?.includes("application/json")) {
         throw new Error(
-          `[DOStateStore] "${method}" request failed with status ${response.status}: Expected JSON response, but got ${response.headers.get("Content-Type")}`,
+          `[CloudflareStateStore] "${method}" request failed with status ${response.status}: Expected JSON response, but got ${response.headers.get("Content-Type")}`,
         );
       }
       const json = (await response.json()) as StateStoreProxy.Response<
@@ -79,7 +79,7 @@ export class DOStateStore extends StateStoreProxy {
       >;
       if (!json.success) {
         throw new Error(
-          `[DOStateStore] "${method}" request failed with status ${response.status}: ${json.error}`,
+          `[CloudflareStateStore] "${method}" request failed with status ${response.status}: ${json.error}`,
         );
       }
       return json.result;
@@ -87,8 +87,8 @@ export class DOStateStore extends StateStoreProxy {
   }
 }
 
-const provision = memoize(async (options: DOStateStoreOptions) => {
-  const scriptName = options.scriptName ?? "alchemy-state";
+const provision = memoize(async (options: CloudflareStateStoreOptions) => {
+  const scriptName = options.scriptName ?? "alchemy-state-service";
   const token =
     options.stateToken ??
     (await alchemy.secret.env(
@@ -99,12 +99,14 @@ const provision = memoize(async (options: DOStateStoreOptions) => {
 
   const api = await createCloudflareApi(options);
   const [bundle, settings, subdomain] = await Promise.all([
-    getInternalWorkerBundle("do-state-store"),
+    getInternalWorkerBundle("cloudflare-state-store"),
     getWorkerSettings(api, scriptName),
     getWorkerSubdomain(api, scriptName),
   ]);
   if (!settings || !settings.tags.includes(bundle.tag) || options.forceUpdate) {
-    logger.log(`[DOStateStore] ${settings ? "Updating" : "Creating"}...`);
+    logger.log(
+      `[CloudflareStateStore] ${settings ? "Updating" : "Creating"}...`,
+    );
     await putWorker(api, {
       workerName: scriptName,
       compatibilityDate: BUILD_DATE,
@@ -149,11 +151,11 @@ async function pollUntilReady(fn: () => Promise<Response>) {
     }
     if (res.status === 401) {
       throw new Error(
-        "[DOStateStore] The token is invalid. Please check your ALCHEMY_STATE_TOKEN environment variable, or set `forceUpdate: true` in the DOStateStore constructor to overwrite the current token.",
+        "[CloudflareStateStore] The token is invalid. Please check your ALCHEMY_STATE_TOKEN environment variable, or set `forceUpdate: true` in the CloudflareStateStore constructor to overwrite the current token.",
       );
     }
     if (!last) {
-      logger.log("[DOStateStore] Waiting for deployment...");
+      logger.log("[CloudflareStateStore] Waiting for deployment...");
     }
     last = res;
     // Exponential backoff with jitter
@@ -163,6 +165,6 @@ async function pollUntilReady(fn: () => Promise<Response>) {
     delay = Math.min(delay, 10000); // Cap at 10 seconds
   }
   throw new Error(
-    `[DOStateStore] Failed to reach state store: ${last?.status} ${last?.statusText}`,
+    `[CloudflareStateStore] Failed to reach state store: ${last?.status} ${last?.statusText}`,
   );
 }
