+++ alchemy/src/state/index.ts
@@ -1,5 +1,5 @@
+export * from "./cloudflare-state-store.ts";
 export * from "./d1-state-store.ts";
-export * from "./do-state-store.ts";
 export * from "./file-system-state-store.ts";
 export * from "./r2-rest-state-store.ts";
 export * from "./sqlite-state-store.ts";
