+++ alchemy/test/util.ts
@@ -1,8 +1,8 @@
 import fs from "node:fs/promises";
 import os from "node:os";
 import type { AlchemyOptions } from "../src/alchemy.ts";
+import { CloudflareStateStore } from "../src/state/cloudflare-state-store.ts";
 import { D1StateStore } from "../src/state/d1-state-store.ts";
-import { DOStateStore } from "../src/state/do-state-store.ts";
 import { FileSystemStateStore } from "../src/state/file-system-state-store.ts";
 import { SQLiteStateStore } from "../src/state/sqlite-state-store.ts";
 
@@ -44,7 +44,7 @@ export const createTestOptions = (storeType: string, namespace: string) =>
     stateStore: (scope) => {
       switch (storeType) {
         case "do":
-          return new DOStateStore(scope);
+          return new CloudflareStateStore(scope);
         case "fs":
           return new FileSystemStateStore(scope);
         case "d1":
